incusd/storage/drivers: Extract unpackVolume and backupVolume to util
